Extract Active Record related adapter code to a helper module
This allows us to share the code used by AR-related adapters more
easily, without having to move everything to the Base class since not
all adapters are AR-based.
